Move AccountRunResult and mergeStateAfterRun from state.ts to main.ts
state.ts is now focused solely on state file I/O (load/save/path
utilities) and persistent state types. AccountRunResult and
mergeStateAfterRun represent run-time processing logic, so they belong
in main.ts alongside the rest of the run orchestration.
