+++ test/xfetch/main.test.ts
@@ -1,14 +1,16 @@
 import { aroundEach, describe, expect, it } from "vitest";
-import type { PostEntry, RunOptions } from "@/xfetch/main.js";
+import type { AccountRunResult, PostEntry, RunOptions } from "@/xfetch/main.js";
 import {
   buildPostEntry,
   buildRunOutput,
   filterPostsByPattern,
+  mergeStateAfterRun,
   parseArgs,
   parseUsername,
   processAccount,
   sortPostsChronologically,
 } from "@/xfetch/main.js";
+import type { XfetchState } from "@/xfetch/state.js";
 import { emptyState, STATE_VERSION } from "@/xfetch/state.js";
 import type { FetchUserPostsOptions, XClientApi, XPost, XUser } from "@/xfetch/xClient.js";
 
@@ -451,5 +453,61 @@ describe("processAccount", () => {
     });
     expect(processed.posts.map((p) => p.text)).toEqual(["useful content"]);
     expect(processed.accountResult.newLastSeenId).toBe("200");
+  });
+});
+
+// ── mergeStateAfterRun ───────────────────────────────────────
+
+describe("mergeStateAfterRun", () => {
+  const NOW = new Date("2026-04-11T12:00:00.000Z");
+
+  it("updates lastSeenId for successful accounts", () => {
+    const state: XfetchState = {
+      version: STATE_VERSION,
+      accounts: {
+        elonmusk: { lastSeenId: "100", lastCheckedAt: "2026-04-01T00:00:00.000Z" },
+      },
+    };
+    const results: AccountRunResult[] = [{ username: "elonmusk", status: "ok", newLastSeenId: "200" }];
+    const next = mergeStateAfterRun(state, results, NOW);
+    expect(next.accounts.elonmusk).toEqual({
+      lastSeenId: "200",
+      lastCheckedAt: NOW.toISOString(),
+    });
+  });
+
+  it("keeps existing lastSeenId when error occurs", () => {
+    const state: XfetchState = {
+      version: STATE_VERSION,
+      accounts: {
+        elonmusk: { lastSeenId: "100", lastCheckedAt: "2026-04-01T00:00:00.000Z" },
+      },
+    };
+    const results: AccountRunResult[] = [{ username: "elonmusk", status: "error" }];
+    const next = mergeStateAfterRun(state, results, NOW);
+    expect(next.accounts.elonmusk).toEqual({
+      lastSeenId: "100",
+      lastCheckedAt: "2026-04-01T00:00:00.000Z",
+    });
+  });
+
+  it("records baseline_established entries with their newLastSeenId", () => {
+    const state = emptyState();
+    const results: AccountRunResult[] = [
+      { username: "elonmusk", status: "baseline_established", newLastSeenId: "555" },
+    ];
+    const next = mergeStateAfterRun(state, results, NOW);
+    expect(next.accounts.elonmusk).toEqual({
+      lastSeenId: "555",
+      lastCheckedAt: NOW.toISOString(),
+    });
+  });
+
+  it("stores account keys case-insensitively", () => {
+    const state = emptyState();
+    const results: AccountRunResult[] = [{ username: "ElonMusk", status: "ok", newLastSeenId: "1" }];
+    const next = mergeStateAfterRun(state, results, NOW);
+    expect(next.accounts.elonmusk).toBeDefined();
+    expect(next.accounts.ElonMusk).toBeUndefined();
   });
 });
